+++ examples/helper/mod.rs
@@ -1,5 +1,5 @@
 use bevy::prelude::*;
-use bevy_inspector_egui::quick::WorldInspectorPlugin;
+//use bevy_inspector_egui::quick::WorldInspectorPlugin;
 
 #[allow(dead_code)]
 pub mod assets;
@@ -18,7 +18,7 @@ pub struct HelperPlugin;
 
 impl Plugin for HelperPlugin {
     fn build(&self, app: &mut bevy::prelude::App) {
-        app.add_plugins(WorldInspectorPlugin::new());
+        //        app.add_plugins(WorldInspectorPlugin::new());
         app.add_systems(Update, camera::movement);
         app.add_systems(Update, map::rotate);
         #[cfg(feature = "rapier")]