+++ src/physics/mod.rs
@@ -96,9 +96,9 @@ impl<T: TiledPhysicsBackend + Default + 'static + std::marker::Sync + std::marke
     for TiledPhysicsPlugin<T>
 {
     fn build(&self, app: &mut bevy::prelude::App) {
-        app.observe(default_physics_settings::<T>);
-        app.observe(collider_from_object::<T>);
-        app.observe(collider_from_tile::<T>);
+        app.add_observer(default_physics_settings::<T>);
+        app.add_observer(collider_from_object::<T>);
+        app.add_observer(collider_from_tile::<T>);
     }
 }
 
@@ -195,11 +195,7 @@ fn collider_from_tile<
         let world_position = trigger.event().world_position(&map_asset);
         commands
             .entity(trigger.event().tile)
-            .insert(TransformBundle::from_transform(Transform::from_xyz(
-                world_position.x,
-                world_position.y,
-                0.0,
-            )));
+            .insert(Transform::from_xyz(world_position.x, world_position.y, 0.0));
 
         for (object_id, object_data) in collision.object_data().iter().enumerate() {
             if objects_filter.contains(&object_data.name) {